fix(ci): link the split server without db2; split routes TU under the line cap
Two CI-only failures on the feature -> testing PR:

1. Undefined db2_code_index_project_delete in aimee-server: server-owned TUs
   (server/*.c) compile via the generic object rule WITHOUT
   AIMEE_DB2_DISABLED, and the shipped server links no db2 objects — but it
   also keeps NO local lexical index (index_scan_project compiles to a stub
   in build/obj/server/index.o, and the canonical code index lives in
   aimee-kb where the purge fan-out's canonical_index store already deletes
   it). gp_local_index_delete is now a weak 0-success no-op with no db2
   linkage; the seam stays so the delete flow's abort-before-filesystem
   ordering remains unit-testable (tests override it to inject failures).
   Also hoisted destfd_parent to function scope — it was read in the cleanup
   path but declared mid-flow, formally uninitialized on early aborts.

2. line-check: server_http_routes.c hit 2607 lines (hard limit 2500). The
   contiguous git-surface block (clone, clone-org, org-repos, git ops,
   projects, projects/delete, session-dir, editor, credentials, sshkey +
   rh_clone_kb_scan) moved verbatim to a new server/server_http_routes_git.c
   (699 lines; routes.c now 1938), same relocation pattern as
   server_http_config_routes.c — the route table stays in routes.c and
   dispatches via server_http_internal.h declarations.

aimee-server links locally; unit-test-git-project, unit-test-server-http,
unit-test-kb-http-routes, unit-test-kb all pass.
